Move PendingSkimage2Change to top-level to avoid circular import during pytest setup (#8064)
