+++ src/skimage/__init__.py
@@ -48,6 +48,9 @@ util
 
 __version__ = '0.26.1rc0.dev0'
 
+class PendingSkimage2Change(PendingDeprecationWarning):
+    """Warning for upcoming skimage2 behavior changes."""
+    pass
 import lazy_loader as _lazy
 
 __getattr__, *_ = _lazy.attach_stub(__name__, __file__)
