+++ src/skimage/morphology/gray.py
@@ -3,7 +3,7 @@ Grayscale morphological operations
 """
 
 from .misc import default_footprint
-from ..util import PendingSkimage2Change
+from skimage import PendingSkimage2Change
 from .._shared._warnings import warn_external
 
 import skimage2 as ski2
